BButler/requirements.txt
openai
speechrecognition
pystray
Pillow
pypiwin32
tkinter
colorama